% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiner-package.R
\docType{package}
\name{affiner-package}
\alias{affiner}
\alias{affiner-package}
\title{affiner: A Finer Way to Render 3D Illustrated Objects in 'grid' Using Affine Transformations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Dilate, permute, project, reflect, rotate, shear, and translate 2D and 3D points. Supports parallel projections including oblique projections such as the cabinet projection as well as axonometric projections such as the isometric projection. Use 'grid's "affine transformation" feature to render illustrated flat surfaces.
}
\section{Package options}{

The following affiner function arguments may be set globally via \code{\link[base:options]{base::options()}}:
\describe{
\item{affiner_angular_unit}{The default for the \code{unit} argument used by \code{\link[=angle]{angle()}} and \code{\link[=as_angle]{as_angle()}}.
The default for this option is "degrees".}
\item{affiner_grid_unit}{The default for the \code{unit} argument used by \code{\link[=affine_settings]{affine_settings()}}.
The default for this option is "inches".}
}
The following \code{cli} options may also be of interest:
\describe{
\item{cli.unicode}{Whether UTF-8 character support should be assumed.
Along with \code{\link[=l10n_info]{l10n_info()}} used to determine the default of the
\code{use_unicode} argument of \code{\link[=format.angle]{format.angle()}} and \code{\link[=print.angle]{print.angle()}}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://trevorldavis.com/R/affiner/}
  \item Report bugs at \url{https://github.com/trevorld/affiner/issues}
}

}
\author{
\strong{Maintainer}: Trevor L. Davis \email{trevor.l.davis@gmail.com} (\href{https://orcid.org/0000-0001-6341-4639}{ORCID})

Authors:
\itemize{
  \item Trevor L. Davis \email{trevor.l.davis@gmail.com} (\href{https://orcid.org/0000-0001-6341-4639}{ORCID})
}

}
