% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.r
\name{angular_unit}
\alias{angular_unit}
\alias{angular_unit<-}
\title{Get/set angular unit of angle vectors}
\usage{
angular_unit(x)

angular_unit(x) <- value
}
\arguments{
\item{x}{An \code{\link[=angle]{angle()}} vector}

\item{value}{A string of the desired angular unit.  See \code{\link[=angle]{angle()}} for supported strings.}
}
\value{
\code{angular_unit()} returns a string of \code{x}'s angular unit.
}
\description{
\code{angular_unit()} gets/sets the angular unit of \code{\link[=angle]{angle()}} vectors.
}
\examples{
  a <- angle(seq(0, 360, by = 90), "degrees")
  angular_unit(a)
  print(a)
  angular_unit(a) <- "turns"
  angular_unit(a)
  print(a)
}
