% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plane.r
\name{as_point1d}
\alias{as_point1d}
\alias{as_point1d.numeric}
\alias{as_point1d.character}
\alias{as_point1d.Coord1D}
\alias{as_point1d.Point1D}
\title{Cast to Point1D object}
\usage{
as_point1d(...)

\method{as_point1d}{numeric}(a, b, ...)

\method{as_point1d}{character}(x, ...)

\method{as_point1d}{Coord1D}(normal, ...)

\method{as_point1d}{Point1D}(point, ...)
}
\arguments{
\item{...}{Passed to other function such as \code{as_coord2d()}.}

\item{a, b}{Numeric vectors that parameterize the point via the equation \code{a * x + b = 0}.
Note this means that \code{x = -b / a}.}

\item{x}{A (character) vector to be cast to a \link{Point1D} object}

\item{normal}{\link{Coord1D} class object.}

\item{point}{A \link{Point1D} object}
}
\description{
\code{as_point1d()} casts to a \link{Point1D} object.
}
\examples{
p1 <- as_point1d(a = 1, b = 0)
}
