% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine-methods.r
\name{as_transform1d}
\alias{as_transform1d}
\alias{as_transform1d.transform1d}
\alias{as_transform1d.default}
\title{Cast to 1D affine transformation matrix}
\usage{
as_transform1d(x, ...)

\method{as_transform1d}{transform1d}(x, ...)

\method{as_transform1d}{default}(x, ...)
}
\arguments{
\item{x}{An object that can be cast to a}

\item{...}{Further arguments passed to or from other methods}
}
\value{
A \code{\link[=transform1d]{transform1d()}} object
}
\description{
\code{as_transform1d()} casts to a \code{\link[=transform1d]{transform1d()}} affine transformation matrix
}
\examples{
m <- diag(2L)
as_transform1d(m)
}
