% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine-methods.r
\name{as_transform2d}
\alias{as_transform2d}
\alias{as_transform2d.transform2d}
\alias{as_transform2d.default}
\title{Cast to 2D affine transformation matrix}
\usage{
as_transform2d(x, ...)

\method{as_transform2d}{transform2d}(x, ...)

\method{as_transform2d}{default}(x, ...)
}
\arguments{
\item{x}{An object that can be cast to a}

\item{...}{Further arguments passed to or from other methods}
}
\value{
A \code{\link[=transform2d]{transform2d()}} object
}
\description{
\code{as_transform2d()} casts to a \code{\link[=transform2d]{transform2d()}} affine transformation matrix
}
\examples{
m <- diag(3)
as_transform2d(m)
}
