% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine-methods.r
\name{as_transform3d}
\alias{as_transform3d}
\alias{as_transform3d.transform3d}
\alias{as_transform3d.default}
\title{Cast to 3D affine transformation matrix}
\usage{
as_transform3d(x, ...)

\method{as_transform3d}{transform3d}(x, ...)

\method{as_transform3d}{default}(x, ...)
}
\arguments{
\item{x}{An object that can be cast to a}

\item{...}{Further arguments passed to or from other methods}
}
\value{
A \code{\link[=transform3d]{transform3d()}} object
}
\description{
\code{as_transform3d()} casts to a \code{\link[=transform3d]{transform3d()}} affine transformation matrix
}
\examples{
m <- diag(4)
as_transform3d(m)
}
