% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-methods.r
\name{centroid}
\alias{centroid}
\alias{mean.coord2d}
\alias{mean.coord3d}
\title{Compute centroids of coordinates}
\usage{
\method{mean}{coord2d}(x, ...)

\method{mean}{coord3d}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=coord2d]{coord2d()}} object or \code{\link[=coord3d]{coord3d()}} object}

\item{...}{Passed to \code{\link[base:mean]{base::mean()}}}
}
\value{
A \code{\link[=coord2d]{coord2d()}} or \code{\link[=coord3d]{coord3d()}} object of length one
}
\description{
\code{mean()}computes controids for for \code{\link[=coord2d]{coord2d()}} and \code{\link[=coord3d]{coord3d()}} objects
}
\examples{
 p <- coord2d(x = 1:4, y = 1:4)
 mean(p)$print(usage = FALSE)
 (sum(p) / length(p))$print(usage = FALSE) # less efficient alternative

 p <- coord3d(x = 1:4, y = 1:4, z = 1:4)
 mean(p)$print(usage = FALSE)
}
