% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-methods.R
\name{cross_product2d}
\alias{cross_product2d}
\title{Compute 2D vector cross product}
\usage{
cross_product2d(x, y)
}
\arguments{
\item{x}{A \link{Coord2D} class vector.}

\item{y}{A \link{Coord2D} class vector.}
}
\value{
A numeric vector of 2D cross products.
}
\description{
\code{cross_product2d()} computes the 2D cross product (also known as the
perp dot product) of two \link{Coord2D} class vectors.
The 2D cross product of vectors \eqn{(x_1, y_1)} and \eqn{(x_2, y_2)} is the
scalar \eqn{x_1 y_2 - y_1 x_2}.
}
\examples{
x <- as_coord2d(2, 3)
y <- as_coord2d(5, 6)
cross_product2d(x, y)
if (getRversion() >= "4.4.0") {
  crossprod(x, y)
}
}
\seealso{
\code{\link[=dot_product2d]{dot_product2d()}} for the dot product of two \link{Coord2D} vectors.
\code{\link[=cross_product3d]{cross_product3d()}} for the cross product of two \link{Coord3D} vectors.
}
