% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{distance1d}
\alias{distance1d}
\title{1D Euclidean distances}
\usage{
distance1d(x, y)
}
\arguments{
\item{x}{Either a \link{Coord1D} or \link{Point1D} class object}

\item{y}{Either a \link{Coord1D} or \link{Point1D} class object}
}
\description{
\code{distance1d()} computes 1D Euclidean distances.
}
\examples{
p <- as_coord1d(x = 1:4)
distance1d(p, as_coord1d(0))
}
