% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{distance2d}
\alias{distance2d}
\title{2D Euclidean distances}
\usage{
distance2d(x, y)
}
\arguments{
\item{x}{Either a \link{Coord2D} or \link{Line2D} class object}

\item{y}{Either a \link{Coord2D} or \link{Line2D} class object}
}
\description{
\code{distance2d()} computes 2D Euclidean distances.
}
\examples{
  p <- as_coord2d(x = 1:4, y = 1:4)
  distance2d(p, as_coord2d(0, 0))
}
