% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.r
\name{distance3d}
\alias{distance3d}
\title{3D Euclidean distances}
\usage{
distance3d(x, y)
}
\arguments{
\item{x}{Either a \link{Coord3D} class object}

\item{y}{Either a \link{Coord3D} class object}
}
\description{
\code{distance3d()} computes 3D Euclidean distances.
}
\examples{
  p <- as_coord3d(x = 1:4, y = 1:4, z = 1:4)
  distance3d(p, as_coord3d("origin"))
}
