% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle.r
\name{is_angle}
\alias{is_angle}
\title{Test whether an object is an angle vector}
\usage{
is_angle(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_angle()} tests whether an object is an angle vector
}
\examples{
a <- angle(180, "degrees")
is_angle(a)
is_angle(pi)
}
