% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_methods.r
\name{is_coord1d}
\alias{is_coord1d}
\title{Test whether an object has a Coord1D class}
\usage{
is_coord1d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_coord1d()} tests whether an object has a "Coord1D" class
}
\examples{
p <- as_coord1d(x = sample.int(10, 3))
is_coord1d(p)
}
