% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.r
\name{is_coord2d}
\alias{is_coord2d}
\title{Test whether an object is a coord2d object}
\usage{
is_coord2d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_coord2d()} tests whether an object is a coord2d object
}
\examples{
p <- coord2d(x = sample.int(10, 3), y = sample.int(10, 3))
is_coord2d(p)

}
