% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_methods.R
\name{is_ellipse2d}
\alias{is_ellipse2d}
\title{Test whether an object has an Ellipse2D class}
\usage{
is_ellipse2d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_ellipse2d()} tests whether an object has an "Ellipse2D" class
}
\examples{
c1 <- as_ellipse2d(as_coord2d(0.5, 0.5), r = 0.5)
is_ellipse2d(c1)
is_ellipse2d(c1) && all(c1$is_circle)
e1 <- as_ellipse2d(as_coord2d(0, 0), rx = 2, ry = 1)
is_ellipse2d(e1)
is_ellipse2d(e1) && all(e1$is_circle)
}
