% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_coord.r
\name{is_line2d}
\alias{is_line2d}
\title{Test whether an object has a Line2D class}
\usage{
is_line2d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_line2d()} tests whether an object has a "Line2D" class
}
\examples{
p <- as_line2d(a = 1, b = 2, c = 3)
is_line2d(p)

}
