% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_methods.r
\name{is_plane3d}
\alias{is_plane3d}
\title{Test whether an object has a Plane3D class}
\usage{
is_plane3d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_plane3d()} tests whether an object has a "Plane3D" class
}
\examples{
p <- as_plane3d(a = 1, b = 2, c = 3, 4)
is_plane3d(p)
}
