% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_methods.r
\name{is_point1d}
\alias{is_point1d}
\title{Test whether an object has a Point1D class}
\usage{
is_point1d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_point1d()} tests whether an object has a "Point1D" class
}
\examples{
p <- as_point1d(a = 1, b = 5)
is_point1d(p)
}
