% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_methods.R
\name{is_polygon2d}
\alias{is_polygon2d}
\title{Test whether an object has a Polygon2D class}
\usage{
is_polygon2d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_polygon2d()} tests whether an object has a "Polygon2D" class
}
\examples{
p <- as_polygon2d(as_coord2d(x = c(0, 1, 1, 0), y = c(0, 0, 1, 1)))
is_polygon2d(p)
}
