% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_methods.R
\name{is_segment2d}
\alias{is_segment2d}
\title{Test whether an object has a Segment2D class}
\usage{
is_segment2d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_segment2d()} tests whether an object has a "Segment2D" class
}
\examples{
p1 <- as_coord2d(x = c(0, 1), y = c(0, 0))
p2 <- as_coord2d(x = c(1, 1), y = c(0, 1))
s <- as_segment2d(p1, p2 = p2)
is_segment2d(s)
}
