% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine-methods.r
\name{is_transform1d}
\alias{is_transform1d}
\title{Test if 1D affine transformation matrix}
\usage{
is_transform1d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_transform1d()} tests if object is a \code{\link[=transform1d]{transform1d()}} affine transformation matrix
}
\examples{
m <- transform1d(diag(2L))
is_transform1d(m)
is_transform1d(diag(2L))
}
