% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine-methods.r
\name{is_transform2d}
\alias{is_transform2d}
\title{Test if 2D affine transformation matrix}
\usage{
is_transform2d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_transform2d()} tests if object is a \code{\link[=transform2d]{transform2d()}} affine transformation matrix
}
\examples{
m <- transform2d(diag(3L))
is_transform2d(m)
is_transform2d(diag(3L))
}
