% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affine-methods.r
\name{is_transform3d}
\alias{is_transform3d}
\title{Test if 3D affine transformation matrix}
\usage{
is_transform3d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_transform3d()} tests if object is a \code{\link[=transform3d]{transform3d()}} affine transformation matrix
}
\examples{
m <- transform3d(diag(4))
is_transform3d(m)
is_transform3d(diag(4))
}
