% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal.r
\name{normal2d}
\alias{normal2d}
\alias{normal2d.Coord2D}
\alias{normal2d.default}
\alias{normal2d.Line2D}
\title{2D normal vectors}
\usage{
normal2d(x, ...)

\method{normal2d}{Coord2D}(x, normalize = TRUE, ...)

\method{normal2d}{default}(x, normalize = TRUE, ...)

\method{normal2d}{Line2D}(x, ..., normalize = TRUE)
}
\arguments{
\item{x}{Object to compute a 2D normal vector for
such as a \link{Coord2D} object or one coercible
to one using \code{\link[=as_coord2d]{as_coord2d()}}.}

\item{...}{Either ignored or passed to \code{\link[=as_coord2d]{as_coord2d()}}.}

\item{normalize}{If \code{TRUE} coerce to a normalize vector}
}
\value{
A \link{Coord2D} (normal) vector
}
\description{
\code{normal2d()} is an S3 generic that computes a 2D normal vector.
}
\examples{
  p <- as_coord2d(x = 2, y = 3)
  normal2d(p)
  normal2d(p, normalize = FALSE)
}
