% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aabb_polygon2d.R
\name{aabb_polygon2d}
\alias{aabb_polygon2d}
\title{Axis-aligned bounding box polygon}
\usage{
aabb_polygon2d(x, ...)
}
\arguments{
\item{x}{A 2D object with a \code{\link[=range]{range()}} method that returns a \link{Coord2D}
of length two (e.g. \link{Coord2D}, \link{Ellipse2D}, \link{Segment2D}).}

\item{...}{Passed to \code{\link[=range]{range()}}.}
}
\value{
A \link{Polygon2D} object with \verb{$is_convex == TRUE} whose vertices are in
counter-clockwise order.
}
\description{
\code{aabb_polygon2d()} creates an axis-aligned bounding box \link{Polygon2D} object
covering the range of \code{x}.
}
\examples{
# Bounding box for a set of points
pts <- as_coord2d(
  x = runif(20, min = 1, max = 3),
  y = runif(20, min = 1, max = 5)
)
p <- aabb_polygon2d(pts)
p$is_convex
plot(p, border = "black", col = "cyan")
points(pts, pch = 16, cex = 1.5)

# Bounding box for an ellipse
e <- as_ellipse2d(as_coord2d(1, 1), rx = 2, ry = 1, theta = degrees(30))
p2 <- aabb_polygon2d(e)
plot(p2, border = "black", col = "cyan")
lines(e, col = "black", lwd = 2)
}
\seealso{
\code{\link[=rectangle_polygon2d]{rectangle_polygon2d()}} for creating a rectangle by center and
dimensions. \code{\link[=range]{range()}} for the bounding-range methods.
}
