% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiner-package.r
\docType{package}
\name{affiner-package}
\alias{affiner}
\alias{affiner-package}
\title{affiner: A finer way to render 3D objects in 'grid' using affine transformations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Translate, rotate, dilate, and project 2D and 3D points. Use the 'grid's "affine transformation" feature to render textured flat surfaces.
}
\section{Package options}{

The following affiner function arguments may be set globally via \code{\link[base:options]{base::options()}}:
\describe{
\item{affiner_angular_unit}{The default for the \code{unit} argument used by \code{\link[=angle]{angle()}} and \code{\link[=as_angle]{as_angle()}}.
The default for this option is "degrees".}
\item{affiner_grid_unit}{The default for the \code{unit} argument used by \code{\link[=affine_settings]{affine_settings()}}.
The default for this option is "inches".}
\item{affiner_print_usage}{The default for the \code{usage} argument used by \code{\link[=coord2d]{coord2d()}} and \code{\link[=coord3d]{coord3d()}} objects' print method.
The default for this option is \code{FALSE}.}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://trevorldavis.com/R/affiner/}
  \item Report bugs at \url{https://github.com/trevorld/affiner/issues}
}

}
\author{
\strong{Maintainer}: Trevor L Davis \email{trevor.l.davis@gmail.com} (\href{https://orcid.org/0000-0001-6341-4639}{ORCID})

}
