% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiner_options.r
\name{affiner_options}
\alias{affiner_options}
\title{Get affiner options}
\usage{
affiner_options(..., default = FALSE)
}
\arguments{
\item{...}{\code{affiner} package options using \code{name = value}.
The return list will use any of these instead of the current/default values.}

\item{default}{If \code{TRUE} return the default values instead of current values.}
}
\value{
A list of option values.
Note this function \strong{does not} set option values itself but
this list can be passed to \code{\link[=options]{options()}}, \code{\link[withr:with_options]{withr::local_options()}}, or \code{\link[withr:with_options]{withr::with_options()}}.
}
\description{
\code{affiner_options()} returns the \code{affiner} package's global options.
}
\examples{
  affiner_options()

  affiner_options(default = TRUE)

  affiner_options(affiner_angular_unit = "pi-radians")
}
\seealso{
\link{affiner} for a high-level description of relevant global options.
}
