% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.r
\name{coord2d}
\alias{coord2d}
\title{2D coordinate vectors}
\usage{
coord2d(x = numeric(0), y = numeric(0))
}
\arguments{
\item{x}{A numeric vector of x coordinates}

\item{y}{A numeric vector of y coordinates}
}
\description{
\code{coord2d()} creates an \code{\link[R6:R6Class]{R6::R6Class()}} object representing two-dimensional points
represented by Cartesian Coordinates.
}
\examples{
p <- coord2d(x = rnorm(100, 2), y = rnorm(100, 2))
print(p, usage = TRUE, n = 0)
print(p, usage = FALSE, n = 10)
pc <- mean(p) # Centroid
# method chained affine transformation matrices are auto-pre-multiplied
p$
  translate(-pc)$
  shear(x = 1, y = 0)$
  reflect("x-axis")$
  rotate(90, "degrees")

}
