% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord.r
\name{coord3d}
\alias{coord3d}
\title{3D coordinate vectors}
\usage{
coord3d(x = numeric(0), y = numeric(0), z = numeric(0))
}
\arguments{
\item{x}{A numeric vector of x coordinates}

\item{y}{A numeric vector of y coordinates}

\item{z}{A numeric vector of z coordinates}
}
\description{
\code{coord3d()} creates an \code{\link[R6:R6Class]{R6::R6Class()}} object representing three-dimensional points
represented by Cartesian Coordinates.
}
\examples{
p <- coord3d(x = rnorm(100, 2), y = rnorm(100, 2), z = rnorm(100, 2))
print(p, usage = TRUE, n = 0)
print(p, usage = FALSE, n = 10)
pc <- mean(p) # Centroid

}
