% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_methods.r
\name{is_coord2d}
\alias{is_coord2d}
\title{Test whether an object has a Coord2D class}
\usage{
is_coord2d(x)
}
\arguments{
\item{x}{An object}
}
\value{
A logical value
}
\description{
\code{is_coord2d()} tests whether an object has a "Coord2D" class
}
\examples{
p <- as_coord2d(x = sample.int(10, 3), y = sample.int(10, 3))
is_coord2d(p)
}
