% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord-methods.r
\name{norm}
\alias{norm}
\alias{abs.coord2d}
\alias{abs.coord3d}
\title{Compute Euclidean norm}
\usage{
\method{abs}{coord2d}(x)

\method{abs}{coord3d}(x)
}
\arguments{
\item{x}{A \code{\link[=coord2d]{coord2d()}} object or \code{\link[=coord2d]{coord2d()}} object.}
}
\value{
A numeric vector
}
\description{
\code{abs()} computes the Euclidean norm for \code{\link[=coord2d]{coord2d()}} objects and \code{\link[=coord3d]{coord3d()}} objects.
}
\examples{
  z <- complex(real = 1:4, imaginary = 1:4)
  p <- as_coord2d(z)
  abs(p) # Euclidean norm
  sqrt(p * p) # Less efficient way to calculate same numbers

  # In {base} R `abs()` calculates Euclidean norm of complex numbers
  all.equal(abs(p), abs(z))
  all.equal(Mod(p), Mod(z))

  p3 <- coord3d(x = 1:4, y = 1:4, z = 1:4)
  abs(p3)
}
