% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_plane.r
\name{Plane3D}
\alias{Plane3D}
\title{3D planes R6 Class}
\description{
\code{Plane3D} is an \code{\link[R6:R6Class]{R6::R6Class()}} object representing three-dimensional planes.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{a}}{Numeric vector that parameterizes the plane via the equation \code{a * x + b * y + c * z + d = 0}.}

\item{\code{b}}{Numeric vector that parameterizes the plane via the equation \code{a * x + b * y + c * z + d = 0}.}

\item{\code{c}}{Numeric vector that parameterizes the plane via the equation \code{a * x + b * y + c * z + d = 0}.}

\item{\code{d}}{Numeric vector that parameterizes the plane via the equation \code{a * x + b * y + c * z + d = 0}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Plane3D-new}{\code{Plane3D$new()}}
\item \href{#method-Plane3D-print}{\code{Plane3D$print()}}
\item \href{#method-Plane3D-clone}{\code{Plane3D$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plane3D-new"></a>}}
\if{latex}{\out{\hypertarget{method-Plane3D-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plane3D$new(a, b, c, d)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{a}}{Numeric vector that parameterizes the plane via the equation \code{a * x + b * y + c * z + d = 0}.}

\item{\code{b}}{Numeric vector that parameterizes the plane via the equation \code{a * x + b * y + c * z + d = 0}.}

\item{\code{c}}{Numeric vector that parameterizes the plane via the equation \code{a * x + b * y + c * z + d = 0}.}

\item{\code{d}}{Numeric vector that parameterizes the plane via the equation \code{a * x + b * y + c * z + d = 0}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plane3D-print"></a>}}
\if{latex}{\out{\hypertarget{method-Plane3D-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plane3D$print(n = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of lines to print.  If \code{NULL} print all of them.}

\item{\code{...}}{Passed to \code{\link[=format.default]{format.default()}}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Plane3D-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Plane3D-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plane3D$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
